requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
supabase==2.3.0
pydub==0.25.1
python-multipart==0.0.6
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
jinja2==3.1.2
Flask-Limiter==3.5.0
